Fix for weird "Promise was collected" error (Masked by Execution context was destroyed)
When doing a `mount` on a fairly large component with a large enough
props payload, I'm randomly getting an "Error: Execution context was
destroyed, most likely because of a navigation.". Which by looking at
`DEBUG` logs, seems to happen due to a "Promise was collected" error
from the CDP `Runtime.callFunctionOn` request.

There was no navigation involved at this point. Rather it seems like the
actual `Promise` that is invoked by `Runtime.callFunctionOn` is getting
somehow collected by the GC, which is likely trigged by either the size
of the component bundle or whatever it does with the props that creates
enough garbage.

This is fairly difficult to reproduce outside of our proprietary
project, but I might be able to show this happening on a video call if
you wish to (And my company agrees). Though will let you know if I
manage to make a standalone reproduction.

It seems that just wrapping the return value in `Promise.resolve` made
this go away, somehow making the VM hold the Promise alive long enough
instead of collecting while it's not done yet.