devops: make sure uploaded reports always contain directory (#29080)
* Before the change if if uploaded artifact path contained just one file
it would be [archived
directly](https://github.com/actions/upload-artifact/blob/694cdabd8bdb0f10b2cea11669e1bf5453eed0a6/src/search.ts#L142-L151),
but if there is a sister folder it would be an archive of the entire
dir. According to the action docs "If a wildcard pattern is used, the
path hierarchy will be preserved after the first wildcard pattern", so
hopefully it means that the directory is always compressed even if there
is a single file in it.
* Use `overwrite: true` to have only one artifact with pull request
number.