fix(recorder): rework extendInjectedScript
Previously, new `Recorder` instance was given an existing `InjectedScript`.
However, we built a separate source for `InjectedScript` vs `Recorder`,
and both bundles contain their own copy of all helper modules, e.g. `roleUtils`.

This resulted in two copies of helper modules, which is troublesome
for any module-level globals like a top-level cache. Depending on
whether `Recorder` or `InjectedScript` called into the helper, they
would access the different value of a module global, which lead to bugs.

New approach creates a separate `InjectedScript` instance for the
`Recorder`, with its own universe of modules. This way we may end up
with two `InjectedScript` instances, but there should be no issues
with that since they are isolated.

Alternative approach considered was to bundle everything together
all the time. This approach increases the size of `InjectedScriptSource`
from 240k to 280k, which is acceptible. However, we may not be able
to include things like `CodeMirror` in the recorder anymore.