fix(recorder): disallow external imports
Previously, new `Recorder` instance was given an existing `InjectedScript`.
However, we built a separate source for `InjectedScript` vs `Recorder`,
and both bundles contain their own copy of all helper modules, e.g. `roleUtils`.

This resulted in two copies of helper modules, which is troublesome
for any module-level globals like a top-level cache. Depending on
whether `Recorder` or `InjectedScript` called into the helper, they
would access the different value of a module global, which lead to bugs.

To prevent this, we force any external dependencies to be imported
through the `InjectedScript.utils`.