chore: do not set metadata.error for expect failure results
The metadata.error change was brought back in https://github.com/microsoft/playwright/pull/29271and it broke java port
as we could have error and result set simulteniously. This PR moves the logic to the trace recorder instead
and keeps the protocol contract clear that either error or result is present, but not both.