use addEventListener instead of onmessage in portTransport.ts
When running Playwright 1.42.1 on Node 18.18.2 under `bazel run`, we found an edge case where `ui mode` would start but the list of tests would never render. When debugging with `--inspect-brk` we were unable to reproduce the issue. Upon logging out where the messages were getting lost from the websocket server, we found this line to be the offending line. Switching `onmessage` to `addEventListener('message'` solved the issue for us and the messages were sent over websocket just fine.

Unfortunately, we were unable to reproduce this issue with just Node in a browser, so we believe the core of the issue is in playwright itself. We aren't sure why this fixes the issue in our specific case, but we wanted to upstream it so that the fix is known.

Signed-off-by: David <david.aghassi@gmail.com>