feat(test runner): show multiple errors, at most one per stage
Previously, there was at most one "hard error", as opposite to
multiple "soft errors". This was done to preserve the historic
behavior at the time of introducing multiple `TestInfo.errors`.

With this change, every user callback that is executed
`withRunnable()` can throw an error and/or timeout, and both of
these will end up in `TestInfo.errors`.

Additionally, there is at most one "unhandled exception" error,
to avoid flooding the report with mass failures.