chore: deprecate/remove noWaitAfter from some actions
The following actions keep `noWaitAfter` option:
- `click`
- `selectOption`
- `press`

All other actions that used to have `noWaitAfter` now
only behave like it was set to true, not waiting for
follow-up navigations. In the docs, this option is marked
as completely ignored.

A small logic change was made to compensate for this behavior:
when waiting for the `hitTargetInterceptor`, we now race it
against navigations to avoid stalling when navigation stalls.
Previously, waiting for the interceptor was disabled when
`noWaitAfter` was passed, and since it's impossible to pass
this option now, we mitigate with never stalling instead.