fix(ui): only populate settings on main page
before, the init script was executed for every loaded iframe and was resetting `localStorage` everytime. It didn't show because the UI had its own copy of the settings via useState() - but now that it's reading directly from local storage, it shows.

The fix is to ensure that the script isn't executed in iframes.