feat: wait for pending navigation to resolve before many actions
This includes all actions that perform locator handler check.

Note this makes it impossible to interact with the page while
a main frame navigation is ongoing. This was already the case
for Chromium, but now WebKit and Firefox align with it.

Setting `PLAYWRIGHT_SKIP_NAVIGATION_CHECK` environment variable
disables this behavior.