Update docs/src/locators.md
The section for "Matching one of the two alternative locators" mentioned "matching all of the alternatives". However, since we are using or() operator, it should match "any one of the alternatives" to locate the element.

As per the suggestion in the review comment, instead of "any one of the alternatives" it was decided to go ahead with "any one or both of the alternatives" to be as clear and explicit as possible.

Co-authored-by: Dmitry Gozman <dgozman@gmail.com>
Signed-off-by: Anand M Cherian <63868951+Anand-M-Cherian@users.noreply.github.com>