fix(chromium): race between oopif attach and context clear
There is a race between:
- `Runtime.executionContextsCleared` coming in the OOPIF.
- `Page.frameNavigated`/`Target.attachedToTarget` when frame goes local->remote.

As a result, we can have stale execution contexts from the previous non-OOPIF
frame that does not exist anymore, instead of new contexts in the OOPIF.

This causes things like recorder trying to initialize in the wrong context and fail.

Now that we clear contexts upon local->remote switch, recorder waits for
the new context to be created an successfully initializes.