fix(test runner): esm loader in old Node versions
Arguments of the `register()` call have changed over time,
in particular the `parentURL` was not a part of options.

We do not need to pass it, because the first argument `specifier`
is already an absolute file url, so we do not need a base url.

Verified this fix works on Node versions 20.11, 20.6, 18.17 and 22.13.