chore: builtins
This introduces a "script to evaluate on new document" that
saves certain builtin objects like `Map` into an unguessable
global property. Later on, every injected script will try to
use the saved builtins.

The idea is to workaround pages that mess around with certain
objects like `Set` or `Map` and break Playwright's injected code.

All injected source files are now subject to an eslint rule
that forces the use of `Builtins.Map` instead of `Map` or similar.
Many utility functions got a `builtins` argument to be able to
construct non-overridden builtins.

Pros:
- Playwright should work on a broader set of web pages.
- Over time, we can cover more builtin classes/functions.

Cons:
- Passing around `builtins` object in the codebase.
- Extra eslint rule.
- A tiny but mandatory script evaluated on each page load.