chore: reduce scrolling during clicks
Existing logic always tries to scroll to a different position after
clicking at a wrong hit target. This is done to scroll out from
under a sticky/fixed header.

However, it does not make sense to scroll if the sticky/fixed header
is not a parent of the wrong hit target element, or is a common parent
of the hit target and the actual target element.

This change restricts the scrolling logic to be only applied when
there is a sticky/fixed element in the DOM chain between the hit target
and its least common ancestor with the target element.