fix(snapshotter): mismatching snapshot index
When we fail to capture snapshot in a frame, for example when
the frame is blocked on a sync javascript dialog, we can end
up in a situation where:
- snapshotter injected script has incremented the snapshotNumber
- meanwhile, snapshotter in node has dropped a snapshot

Now, all subsequent snapshots in the same frame will have
mismatching snapshot numbers between injected and node sides.
This means all "reference nodes" will be wrong and re-combined
snapshots will be incorrect.

The fix is to force a reset after a failed snapshot,
so that future snapshots do not reference old nodes.