feat(fixtures): support `box: 'self'`
Fixture marked as `box: 'self'` will be invisible in the report.
However, all the child steps will be inlined instead of the boxed
fixture - visible in the report as children of the boxed fixture's
parent step.

No changes in the trace viewer, all boxed fixtures are hidden by
default, and visible when "configuration" action filter is checked.