fix(test runner): unhandled rejection in test.fail should not skip other tests
We stop the worker inside the `WorkerMain` upon unhandled rejection in
`test.fail`. However, dispatcher does not know about that and sends a new
batch of tests to the same worker, that ultimately end up as ignored
because the worker has stopped.

An explicit flag in the `DonePayload` fixes the problem.