fix(clock): make sure runTo is atomic
Since `runTo` is asynchronous, we might end up in a reentrant call,
for example when `updateRealTimeTimer` timeout happens between two
tasks in `runTo`. This yields to breaking all invariants and results
in `now` being reset to some old value.

To prevent this, introduce a semaphore and only allow a single `runTo`
at a time.

Additionally, ensure we never rewind time when running our tests.